🛠️ fix(client): Corrected URL path in PostConfig.js
Fixed a typo in the URL path for the drawio assets in PostConfig.js. The path was incorrectly prefixed with a slash, which has been removed to ensure the correct relative path is used.

🛠️ fix(scripts): Adjusted createZip condition in copy_and_bundle_build.js

Modified the condition for creating a zip file in copy_and_bundle_build.js to ensure it only creates a zip when the output directory is 'dist'. This prevents unnecessary zip creation in other output directories.